+++ CHANGELOG.md
@@ -7,3 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Task Executor interface for Furucombo Smart Wallet.
+- DSProxy auth action
+- Furucombo action auth action
+- Trevi action
+- Wallet action